Implement handleFavorite
Created new props entityType, entityId, isFavInit, showFavIcon.  These allow the Cards props to be used every the card is used on the app.

Added Heart icon that use post request to database when isFavorite is true and uses Delete request when isFavorite is false.   Icon lights up red when isFavorite.